Added documentation to "action" crate
