+++ src/action/mod.rs
@@ -5,7 +5,6 @@ mod publisher;
 mod routine;
 mod action;
 mod actions;
-mod types;
 
 pub use command::*;
 pub use handler::*;
@@ -13,4 +12,3 @@ pub use publisher::*;
 pub use routine::*;
 pub use action::*;
 pub use actions::*;
-pub use types::*;
