+++ src/action/routine.rs
@@ -6,12 +6,14 @@ use crate::storage::{Chronicle, Log};
 use chrono::{DateTime, Utc};
 use std::sync::{Arc, Mutex, Weak};
 
-/// A `Command` that should be executed at a scheduled time *outside* of the normal event loop.
+/// A [`Command`] that should be executed at a scheduled time *outside* of the normal event loop.
 ///
 /// A weak reference to originating log is maintained so that logging of events is automatically
 /// handled.
 ///
-/// Typically these should exclusively be `Output` events, such as completing a time bound operation.
+/// Typically, a [`Routine`] should be intended for output events should be generated by [`IOEvent`]s
+/// that are generated by input data. Such events would be time bound operations such as toggling a
+/// device to only keep it active (or or inactive) for a short period of time.
 ///
 /// # Example
 /// The primary use case is turning off a pump or other output after a predetermined period of time.
